requirements.txt
django >= 2
djangorestframework
gunicorn
psycopg2
django-cors-headers
django-extensions